dla marcina

+++ piwerko-backend/piwerko-backend/Controllers/BreweryController.cs
@@ -128,8 +128,9 @@ namespace Piwerko.Api.Controllers
         }
 
         [HttpPost("delete")]
-        public IActionResult Delete([FromBody]JObject data)
+        public IActionResult Delete([FromBody]JObject data, [FromHeader(Name = "user_id")] int user_id)
         {
+            if (!_userService.isAdmin(user_id)) return BadRequest("nie jestes adminem");
             int index = data["id"].ToObject<Int32>();
             if (_breweryService.Delete(index)) return Ok();
             return NotFound("brak browaru o danym id");
